(chore) refactor NewItemRow
